--- src/main/java/com/ap4j/bma/apitest_kjh/Apitest_kjh_RestController.java
@@ -1,47 +1,0 @@
-package com.ap4j.bma.apitest_kjh;
-
-import com.fasterxml.jackson.databind.util.JSONPObject;
-import org.json.JSONObject;
-import org.json.XML;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.io.BufferedInputStream;
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-
-@RestController
-public class Apitest_kjh_RestController {
-
-    @PostMapping("/map")
-    public String mapPage(){
-        StringBuffer result = new StringBuffer();
-        String jsonPrintString = null;
-        try {
-            String apiUrl = "http://openapi.molit.go.kr:8081/OpenAPI_ToolInstallPackage/service/rest/RTMSOBJSvc/getRTMSDataSvcAptTrade?_wadl&type=xml?" +
-                    "serviceKey=HIVZzhQpW%2BXuIwc8TdHtjjITL8zkdjbLHMe80yY8Ub21YCfA%2BvEK2RUqTXp7OycPYmwRrFEsQisnLuJrSY24JQ%3D%3D" +
-                    "&numOfRows=10" +
-                    "&pageNo=4";
-            URL url = new URL(apiUrl);
-            HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
-            urlConnection.connect();
-            BufferedInputStream bufferedInputStream = new BufferedInputStream(urlConnection.getInputStream());
-            BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(bufferedInputStream, "UTF-8"));
-            String returnLine;
-            while((returnLine = bufferedReader.readLine()) != null) {
-                result.append(returnLine);
-            }
-
-            JSONObject jsonObject = XML.toJSONObject(result.toString());
-            jsonPrintString = jsonObject.toString();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-
-        return jsonPrintString;
-    }
-
-}
