+++ src/main/resources/templates/busan/busanAjax.html
@@ -7,6 +7,12 @@
 <body>
 <div id="map" style="width:500px;height:400px;"></div>
 <script type="text/javascript" src="//dapi.kakao.com/v2/maps/sdk.js?appkey=c913c99f4422dbc35dcae0255c68055f"></script>
-<script th:src="@{/apitest_kjh/js/map.js}"></script>
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
+<script th:src="@{/js/busanAjax.js}"></script>
+<p>
+
+    [[${data}]]
+
+</p>
 </body>
 </html>
